+++ frontend/src/services/api/usersRequests.ts
@@ -1,0 +1,6 @@
+import { Users } from "./usersTypes";
+
+export const fetchUsers = async () => {
+	const response = await fetch("http://localhost:4000/users");
+	return (await response.json()) as Users[];
+};
